refactor and modularise code
