+++ src/service_index/logic/mod.rs
@@ -1,1 +1,2 @@
 pub mod service_canisters_logic;
+pub mod proposal_logic;
