+++ src/service_index/methods/mod.rs
@@ -1,1 +1,2 @@
 pub mod service_canister_methods;
+pub mod proposal_methods;
