+++ src/service_index/misc/mod.rs
@@ -1,1 +1,2 @@
 pub mod utils;
+pub mod ic_utils;
